requirements.txt
asgiref==3.11.0
Django==6.0
django-restframework==0.0.1
djangorestframework==3.16.1
sqlparse==0.5.4
